Fix booking page: quote price integration, insurance selection, payment flow, and error handling
- Fixed booking page to use quote price as base and add insurance on top
- Updated price calculation to use quoteBasePrice from quote data
- Fixed CSS selector error for service selection section
- Fixed Google Maps initMap callback error by defining function globally
- Fixed service worker blocking API calls by skipping cache for /api/ routes
- Fixed Paystack invalid key error with fallback to mock mode
- Fixed app.js interference with booking page form submission
- Updated navbar to stay blue in light mode
- Fixed login/register 500 errors by making AuthService lazy and improving error handling
- Added receiver SMS notifications for booking confirmations
- Improved booking form auto-population from quote and logged-in user data

+++ src/main/java/com/reliablecarriers/Reliable/Carriers/model/CustomerAddress.java
@@ -3,6 +3,7 @@ package com.reliablecarriers.Reliable.Carriers.model;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Entity
@@ -61,10 +62,10 @@ public class CustomerAddress {
     
     // Google Maps coordinates
     @Column(precision = 10, scale = 8)
-    private Double latitude;
+    private BigDecimal latitude;
     
     @Column(precision = 11, scale = 8)
-    private Double longitude;
+    private BigDecimal longitude;
     
     @Column(length = 255)
     private String placeId; // Google Place ID
@@ -177,19 +178,19 @@ public class CustomerAddress {
         this.contactName = contactName;
     }
     
-    public Double getLatitude() {
+    public BigDecimal getLatitude() {
         return latitude;
     }
     
-    public void setLatitude(Double latitude) {
+    public void setLatitude(BigDecimal latitude) {
         this.latitude = latitude;
     }
     
-    public Double getLongitude() {
+    public BigDecimal getLongitude() {
         return longitude;
     }
     
-    public void setLongitude(Double longitude) {
+    public void setLongitude(BigDecimal longitude) {
         this.longitude = longitude;
     }
